+++ lib/torch.rb
@@ -174,10 +174,8 @@ require "torch/nn/init"
 
 # utils
 require "torch/utils/data/data_loader"
+require "torch/utils/data/dataset"
 require "torch/utils/data/tensor_dataset"
-
-# random
-require "torch/random"
 
 # hub
 require "torch/hub"
@@ -317,12 +315,11 @@ module Torch
     end
 
     def save(obj, f)
-      raise NotImplementedYet unless obj.is_a?(Tensor)
-      File.binwrite(f, _save(obj))
+      File.binwrite(f, _save(to_ivalue(obj)))
     end
 
     def load(f)
-      raise NotImplementedYet
+      to_ruby(_load(File.binread(f)))
     end
 
     # --- begin tensor creation: https://pytorch.org/cppdocs/notes/tensor_creation.html ---
@@ -446,6 +443,100 @@ module Torch
     end
 
     private
+
+    def to_ivalue(obj)
+      case obj
+      when String
+        IValue.from_string(obj)
+      when Integer
+        IValue.from_int(obj)
+      when Tensor
+        IValue.from_tensor(obj)
+      when Float
+        IValue.from_double(obj)
+      when Hash
+        dict = {}
+        obj.each do |k, v|
+          dict[to_ivalue(k)] = to_ivalue(v)
+        end
+        IValue.from_dict(dict)
+      when true, false
+        IValue.from_bool(obj)
+      when nil
+        IValue.new
+      when Array
+        if obj.all? { |v| v.is_a?(Tensor) }
+          IValue.from_list(obj.map { |v| IValue.from_tensor(v) })
+        else
+          raise Error, "Unknown list type"
+        end
+      else
+        raise Error, "Unknown type: #{obj.class.name}"
+      end
+    end
+
+    def to_ruby(ivalue)
+      if ivalue.bool?
+        ivalue.to_bool
+      elsif ivalue.double?
+        ivalue.to_double
+      elsif ivalue.int?
+        ivalue.to_int
+      elsif ivalue.none?
+        nil
+      elsif ivalue.string?
+        ivalue.to_string_ref
+      elsif ivalue.tensor?
+        ivalue.to_tensor
+      elsif ivalue.generic_dict?
+        dict = {}
+        ivalue.to_generic_dict.each do |k, v|
+          dict[to_ruby(k)] = to_ruby(v)
+        end
+        dict
+      elsif ivalue.list?
+        ivalue.to_list.map { |v| to_ruby(v) }
+      else
+        type =
+          if ivalue.capsule?
+            "Capsule"
+          elsif ivalue.custom_class?
+            "CustomClass"
+          elsif ivalue.tuple?
+            "Tuple"
+          elsif ivalue.future?
+            "Future"
+          elsif ivalue.r_ref?
+            "RRef"
+          elsif ivalue.int_list?
+            "IntList"
+          elsif ivalue.double_list?
+            "DoubleList"
+          elsif ivalue.bool_list?
+            "BoolList"
+          elsif ivalue.tensor_list?
+            "TensorList"
+          elsif ivalue.object?
+            "Object"
+          elsif ivalue.module?
+            "Module"
+          elsif ivalue.py_object?
+            "PyObject"
+          elsif ivalue.scalar?
+            "Scalar"
+          elsif ivalue.device?
+            "Device"
+          # elsif ivalue.generator?
+          #   "Generator"
+          elsif ivalue.ptr_type?
+            "PtrType"
+          else
+            "Unknown"
+          end
+
+        raise Error, "Unsupported type: #{type}"
+      end
+    end
 
     def tensor_size(size)
       size.flatten