+++ lib/torch/hub.rb
@@ -4,6 +4,14 @@ module Torch
       def list(github, force_reload: false)
         raise NotImplementedYet
       end
+
+      def download_url_to_file(url)
+        raise NotImplementedYet
+      end
+
+      def load_state_dict_from_url(url)
+        raise NotImplementedYet
+      end
     end
   end
 end