+++ lib/torch/utils/data/dataset.rb
@@ -1,0 +1,8 @@
+module Torch
+  module Utils
+    module Data
+      class Dataset
+      end
+    end
+  end
+end